+++ .eslintrc.json
@@ -26,6 +26,10 @@
 		"semi": [
 			"error",
 			"never"
+		],
+		"eol-last" : [
+			"error",
+			"always"
 		]
 	}
 }
