+++ src/containers/home/index.js
@@ -1,14 +1,73 @@
-import React, { Component } from 'react'
+import React, { Component, PropTypes } from 'react'
 import { translate } from 'react-i18next'
 
+import {
+	timeline
+} from '../../utils/api'
+import { connect } from 'react-redux'
+
 class Home extends Component {
+	constructor (props) {
+		super(props)
+		this.state = {
+			loading: true,
+			timeline : null
+		}
+	}
+
+	componentDidMount() {
+		const { user } = this.props
+
+		if (user) {
+			timeline(user)
+				.then(timeline => {
+					this.setState(() => {
+						return {
+							loading : false,
+							timeline
+						}
+					})
+				})
+		}
+	}
+
 	render() {
 		const { t } = this.props
+		const { loading, timeline } = this.state
+
+		if (loading) {
+			return <div>
+				{t('loading')}
+			</div>
+		}
 
 		return <div>
-            {t('welcome')}
+			{timeline.map((toot, index) => {
+				const content = (toot.reblog && toot.reblog.content) || toot.content
+
+				return <div key={index}>
+					{content}
+					<hr/>
+				</div>
+			})}
 		</div>
 	}
 }
 
-export default translate()(Home)
+Home.propTypes = {
+	user : PropTypes.object.isRequired
+}
+
+const mapStateToProps = (state) => {
+	const { selected, accounts } = state.users
+
+	return {
+		user : accounts[selected]
+	}
+}
+
+export default connect(
+	mapStateToProps
+)(
+	translate()(Home)
+)
